# openapi.yaml
openapi: 3.0.0
info:
  title: WhatsApp Automation Bot API
  description: Production-ready WhatsApp automation bot system with API-like features powered by Playwright
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login to get JWT token
      description: Authenticate using API key to receive JWT token for subsequent requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - apiKey
              properties:
                apiKey:
                  type: string
                  description: Your API key
                  example: "your-secret-api-key-here"
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    description: JWT token
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "admin"
                      role:
                        type: string
                        example: "admin"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/verify:
    post:
      tags:
        - Authentication
      summary: Verify JWT token
      description: Verify if a JWT token is still valid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: JWT token to verify
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      role:
                        type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /messages:
    get:
      tags:
        - Messages
      summary: Get messages with pagination and filtering
      description: Retrieve messages with optional filtering and pagination
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/ChatIdParam'
        - $ref: '#/components/parameters/SenderIdParam'
        - $ref: '#/components/parameters/IsGroupParam'
        - $ref: '#/components/parameters/MessageTypeParam'
        - $ref: '#/components/parameters/StartDateParam'
        - $ref: '#/components/parameters/EndDateParam'
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /messages/stats:
    get:
      tags:
        - Messages
      summary: Get message statistics
      description: Retrieve various statistics about messages
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      total_messages:
                        type: integer
                        example: 1500
                      received_messages:
                        type: integer
                        example: 800
                      sent_messages:
                        type: integer
                        example: 700
                      group_messages:
                        type: integer
                        example: 400
                      personal_messages:
                        type: integer
                        example: 1100
                      unique_chats:
                        type: integer
                        example: 50
                      unique_senders:
                        type: integer
                        example: 30
                      messageTypes:
                        type: array
                        items:
                          type: object
                          properties:
                            message_type:
                              type: string
                            count:
                              type: integer
                      dailyMessages:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              format: date
                            count:
                              type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    ChatIdParam:
      name: chatId
      in: query
      description: Filter by chat ID
      required: false
      schema:
        type: string
    SenderIdParam:
      name: senderId
      in: query
      description: Filter by sender ID
      required: false
      schema:
        type: string
    IsGroupParam:
      name: isGroup
      in: query
      description: Filter by group chat
      required: false
      schema:
        type: boolean
    MessageTypeParam:
      name: messageType
      in: query
      description: Filter by message type
      required: false
      schema:
        type: string
    StartDateParam:
      name: startDate
      in: query
      description: Filter by start date
      required: false
      schema:
        type: string
        format: date-time
    EndDateParam:
      name: endDate
      in: query
      description: Filter by end date
      required: false
      schema:
        type: string
        format: date-time
...
  schemas:
    Message:
      type: object
      properties:
        id:
          type: integer
          example: 1
        message_id:
          type: string
          example: "wamid.HbLqX5Th9t2AAB..."
        chat_id:
          type: string
          example: "1234567890@s.whatsapp.net"
        chat_name:
          type: string
          example: "John Doe"
        sender_id:
          type: string
          example: "1234567890@s.whatsapp.net"
        sender_name:
          type: string
          example: "John Doe"
        message_type:
          type: string
          enum: [text, image, video, audio, document, sticker, contact, location]
          example: "text"
        content:
          type: string
          example: "Hello there!"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00.000Z"
        is_from_me:
          type: boolean
          example: false
        is_group:
          type: boolean
          example: false
        media_url:
          type: string
          format: uri
          example: "https://mmg.whatsapp.net/d/f/..."
        media_type:
          type: string
          example: "image"
        media_size:
          type: integer
          example: 102400
        thumbnail_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00.000Z"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 1000
        totalPages:
          type: integer
          example: 20
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Validation failed"
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Invalid or expired token"
...